backend/requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
dj-database-url==0.4.1
Django==1.8.14
django-classy-tags==0.7.2
django-cms==3.2.5
django-formtools==1.0
django-mptt==0.8.5
django-reversion==1.8.7
django-sekizai==0.9.0
Django-Select2==4.3.2
django-treebeard==4.0.1
djangocms-admin-style==1.2.3
djangocms-attributes-field==0.1.1
djangocms-column==1.6.0
djangocms-file==1.0
djangocms-googlemap==0.5.1
djangocms-inherit==0.2.1
djangocms-installer==0.8.11
djangocms-link==1.8.2
djangocms-picture==0.2.0
djangocms-style==1.7.0
djangocms-teaser==0.2.0
djangocms-text-ckeditor==2.9.4
djangocms-video==1.0.0
html5lib==0.9999999
ipython==5.0.0
ipython-genutils==0.1.0
pathlib2==2.1.0
pexpect==4.2.0
pickleshare==0.7.3
Pillow==3.3.0
prompt-toolkit==1.0.3
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
pytz==2016.6.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
tzlocal==1.2.2
wcwidth==0.1.7
wheel==0.24.0
xmltodict==0.10.2
pyfscache==0.9.12
requests==2.11.1
django-simple-captcha==0.5.3
